Source code cleanup and optimizations (#100)

+++ src/Hangfire.Mongo/MongoWriteOnlyTransaction.cs
@@ -106,6 +106,11 @@ namespace Hangfire.Mongo
 
         public override void IncrementCounter(string key)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -116,6 +121,11 @@ namespace Hangfire.Mongo
 
         public override void IncrementCounter(string key, TimeSpan expireIn)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -127,6 +137,11 @@ namespace Hangfire.Mongo
 
         public override void DecrementCounter(string key)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -137,6 +152,11 @@ namespace Hangfire.Mongo
 
         public override void DecrementCounter(string key, TimeSpan expireIn)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new CounterDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -153,6 +173,11 @@ namespace Hangfire.Mongo
 
         public override void AddToSet(string key, string value, double score)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             var builder = Builders<SetDto>.Update;
             var set = builder.Set(_ => _.Score, score);
             var setTypesOnInsert = builder.SetOnInsert("_t", new[] { nameof(KeyValueDto), nameof(ExpiringKeyValueDto), nameof(SetDto) });
@@ -172,6 +197,11 @@ namespace Hangfire.Mongo
 
         public override void RemoveFromSet(string key, string value)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<SetDto>()
                 .DeleteMany(
@@ -181,6 +211,11 @@ namespace Hangfire.Mongo
 
         public override void InsertToList(string key, string value)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData.InsertOne(new ListDto
             {
                 Id = ObjectId.GenerateNewId(),
@@ -191,6 +226,11 @@ namespace Hangfire.Mongo
 
         public override void RemoveFromList(string key, string value)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<ListDto>()
                 .DeleteMany(
@@ -200,21 +240,27 @@ namespace Hangfire.Mongo
 
         public override void TrimList(string key, int keepStartingFrom, int keepEndingAt)
         {
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x =>
             {
-                int start = keepStartingFrom + 1;
-                int end = keepEndingAt + 1;
-
-                ObjectId[] items = ((IEnumerable<ListDto>)x.StateData.OfType<ListDto>()
-                        .Find(new BsonDocument())
-                        .Project(Builders<ListDto>.Projection.Include(_ => _.Key))
-                        .Project(_ => _)
-                        .ToList())
+                var items = x.StateData
+                    .OfType<ListDto>()
+                    .Find(new BsonDocument())
+                    .Project<KeyValueDto>(Builders<ListDto>.Projection.Include(_ => _.Id))
+                    .ToEnumerable()
+                    .Where((data, i) => (i < keepStartingFrom) || (i > keepEndingAt))
+                    .Select(data => data.Id)
                     .Reverse()
-                    .Select((data, i) => new { Index = i + 1, Data = data.Id })
-                    .Where(_ => ((_.Index >= start) && (_.Index <= end)) == false)
-                    .Select(_ => _.Data)
                     .ToArray();
+
+                if (!items.Any())
+                {
+                    return;
+                }
 
                 x.StateData
                     .OfType<ListDto>()
@@ -226,10 +272,14 @@ namespace Hangfire.Mongo
         public override void SetRangeInHash(string key, IEnumerable<KeyValuePair<string, string>> keyValuePairs)
         {
             if (key == null)
+            {
                 throw new ArgumentNullException(nameof(key));
+            }
 
             if (keyValuePairs == null)
+            {
                 throw new ArgumentNullException(nameof(keyValuePairs));
+            }
 
             var builder = Builders<HashDto>.Update;
             var setTypesOnInsert = builder.SetOnInsert("_t", new[] { nameof(KeyValueDto), nameof(ExpiringKeyValueDto), nameof(HashDto) });
@@ -260,7 +310,9 @@ namespace Hangfire.Mongo
         public override void RemoveHash(string key)
         {
             if (key == null)
+            {
                 throw new ArgumentNullException(nameof(key));
+            }
 
             QueueCommand(x => x.StateData.OfType<HashDto>().DeleteMany(Builders<HashDto>.Filter.Eq(_ => _.Key, key)));
         }
@@ -279,15 +331,16 @@ namespace Hangfire.Mongo
         }
 
 
-
-        //New methods to support Hangfire pro feature - batches.
-
-
+        // New methods to support Hangfire pro feature - batches.
 
 
         public override void ExpireSet(string key, TimeSpan expireIn)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x
                 .StateData
                 .OfType<SetDto>()
@@ -297,7 +350,11 @@ namespace Hangfire.Mongo
 
         public override void ExpireList(string key, TimeSpan expireIn)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<ListDto>()
                 .UpdateMany(Builders<ListDto>.Filter.Eq(_ => _.Key, key),
@@ -306,7 +363,11 @@ namespace Hangfire.Mongo
 
         public override void ExpireHash(string key, TimeSpan expireIn)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<HashDto>()
                 .UpdateMany(Builders<HashDto>.Filter.Eq(_ => _.Key, key),
@@ -315,7 +376,11 @@ namespace Hangfire.Mongo
 
         public override void PersistSet(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<SetDto>()
                 .UpdateMany(Builders<SetDto>.Filter.Eq(_ => _.Key, key),
@@ -324,7 +389,11 @@ namespace Hangfire.Mongo
 
         public override void PersistList(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<ListDto>()
                 .UpdateMany(Builders<ListDto>.Filter.Eq(_ => _.Key, key),
@@ -333,7 +402,11 @@ namespace Hangfire.Mongo
 
         public override void PersistHash(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+
             QueueCommand(x => x.StateData
                 .OfType<HashDto>()
                 .UpdateMany(Builders<HashDto>.Filter.Eq(_ => _.Key, key),
@@ -342,8 +415,14 @@ namespace Hangfire.Mongo
 
         public override void AddRangeToSet(string key, IList<string> items)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
-            if (items == null) throw new ArgumentNullException(nameof(items));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
+            if (items == null)
+            {
+                throw new ArgumentNullException(nameof(items));
+            }
             var builder = Builders<SetDto>.Update;
 
 
@@ -373,7 +452,10 @@ namespace Hangfire.Mongo
 
         public override void RemoveSet(string key)
         {
-            if (key == null) throw new ArgumentNullException(nameof(key));
+            if (key == null)
+            {
+                throw new ArgumentNullException(nameof(key));
+            }
             QueueCommand(x => x.StateData
                 .OfType<SetDto>()
                 .DeleteMany(Builders<SetDto>.Filter.Eq(_ => _.Key, key)));
